Add regex-aware snapshot updates
Seems like it works ok.  I had something 'fancy' in mind, involving
tailoring the diff to keep the regexes, but my approach didn't work.

As I think about it more, I do think there's a way to get that correct,
but this is actually more robust, as well as simpler.