Limit regex match to one line, accept identicals
This commit makes two small tweaks, each likely to improve the behavior
of the library.

The first is to prevent any newlines in a regex match.  They're invalid
in a regex and would indicate an overbroad matched region.

The second is to accept a diff if the matched regex is identical to the
text it should 'match'.  Doing otherwise would be creating issues where
there were none, since the point is to match the old and the new.